+++ src/loadVariables.ts
@@ -1,20 +1,19 @@
 import { ApiError } from "./ApiError";
 import { ObjectType } from "./interfaces";
-import { isOfType } from "./utils/isOfType";
 
+// to be removed
 export interface DefaultValues {
-  // to remove
   [name: string]: any;
 }
 
-type ObjectDefinition = { default?: any; parser?: Function };
+type ObjectDefinition = { default?: any; parser?: (v: any) => any };
 type PrimitiveDefinition = boolean | number | string | symbol;
 
 export interface VariableDefinitions {
   [key: string]: PrimitiveDefinition | ObjectDefinition;
 }
 
-type GetSimpleType<T> = T;
+type GetSimpleType<T> = T extends boolean ? boolean : T;
 type GetNestedType<T, K extends keyof T> = GetSimpleType<T[K]>;
 type GetType<T> = T extends ObjectType
   ? GetNestedType<T, "default">
@@ -30,7 +29,7 @@ function getDefinition(
   return typeof def != "object" ? { default: def } : def;
 }
 
-export const loadEnvVars = <T extends VariableDefinitions>(vars: T) => {
+export const loadVariables = <T extends VariableDefinitions>(vars: T) => {
   const error = new ApiError({
     message: "Invalid Environment Variables",
     statusCode: 500,
@@ -50,7 +49,7 @@ export const loadEnvVars = <T extends VariableDefinitions>(vars: T) => {
 
       const { default: _default, parser } = getDefinition(definition);
 
-      if (parser && !isOfType(parser, "function")) {
+      if (parser && typeof parser != "function") {
         error.add(name, "Invalid parser");
         isInValid = true;
       }
@@ -73,16 +72,13 @@ export const loadEnvVars = <T extends VariableDefinitions>(vars: T) => {
   return parsedVars;
 };
 
-const res = loadEnvVars({
-  age: 20,
-  isAdmin: false,
-  name: "James",
-  revnameiew: { default: "baba" },
-  review: { parser: () => {}, default: 20 },
-});
+const env = {
+  DB_NAME: "test-deb",
+  ETA: 20,
+  IS_DEBUG_OPEN: false,
+  MAX_TIME_TO_CANCEL: { parser: (v: any) => v, default: 25 },
+  TO_REJECT: { default: ["apple", "potato"] },
+};
 
-res.age;
-res.isAdmin;
-res.name;
-res.revnameiew;
-res.review;
+const { DB_NAME, ETA, IS_DEBUG_OPEN, MAX_TIME_TO_CANCEL, TO_REJECT } =
+  loadVariables(env);