+++ src/utils/isOfType.ts
@@ -1,1 +1,0 @@
-export const isOfType = (v: any, _type: string = "") => typeof v == _type;